backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv

# correctos:
email-validator
passlib==1.7.4
bcrypt==4.0.1
python-jose
python-multipart
